WIP perf tests: move perf-utils.ts to the performance folder
